test(gateway): put the fold's correctness facts in the suite that runs on every gate
Main split the Gateway tests: the pure ones moved to CcDirector.Gateway.UnitTests, which
runs in the two-minute default gate, and the host-bound remainder is PARKED behind
-Parked along with Core.Tests. Both of this work item's files landed in the parked half by
default, so the default gate would have reported green without executing either of them -
this mission's own false-green shape, arriving as the default behaviour rather than as an
accident.

Split by what each fact actually needs, rather than leaving them where they fell.

The five CORRECTNESS facts move to the fast suite and now run on every gate: that the
batched read answers exactly what the three per-session reads answered, that a deferred
hold is not an absent one, that the snapshot and the per-session readers cannot disagree,
that the read asks only about the sessions being folded, and that it stays inside its
tenant. None of them constructs a host, binds a port, or touches process-global state.

The four READ-COUNT facts stay parked, and the reason is mechanical rather than cautious:
each asserts a delta on LoadTestMetrics.snoozeDbReads, which is process-global, and the
fast assembly runs four collections at once beside a dozen fold tests that increment that
same counter. An exact delta measured there would quietly absorb another test's reads,
making this item's headline proof flaky. The locked suite disables parallelism, which is
what makes the delta mean anything - the same rule the split applied to the classes that
mutate process-wide environment variables.

The sweep guard tests stay parked too: they boot a real GatewayHost and bind a port.

The count comparison against test-baseline.md is retired for the two split projects, whose
totals moved by about 2,750 for reasons unrelated to this branch; comparing against them
would produce an alarming fall that means nothing. What replaces it is stronger after a
layout change: the thirteen facts are named, and must appear in the TRX as EXECUTED and
passed. Every report from here states which invocation was run, because green now means
two different things depending on the flag.
